r-chess-prototype/requirements.txt
nicegui 
python-chess 
uvicorn